+++ app/routes/index.js
@@ -66,36 +66,32 @@ router.use(function(req, res, next) {
     res.render('get_fragment_hack.ejs');
 });
 
-// TODO[gio]:: this route is unused, since / gets handled in login, maybe delete?
-router.get('/', function(req, res) {
-    res.type('html');
-    res.redirect('/home');
-});
-
 router.get('/process-order/:orderId', function(req, res, next) {
     res.type('html');
     // render view to show progress
     res.render('processing-order');
 });
 
-router.get('/home', function(req, res) {
+router.get('/projects', function(req, res) {
     res.type('html');
     res.render('home', {
-        title: "Home",
+        title: "Projects",
         user: req.session.user,
-        state: 'home.my',
+        auth0UniversalLoginUrl: auth0Service.universalLoginUrl,
+        state: 'projects',
         inject_data: {
             mapbox_access_token: config('mapbox-api').accessToken
         },
     });
 });
 
-router.get('/home/all', function(req, res) {
+router.get('/', function(req, res) {
     res.type('html');
     res.render('home', {
-        title: "Home",
+        title: "Projects",
         user: req.session.user,
-        state: 'home.all',
+        auth0UniversalLoginUrl: auth0Service.universalLoginUrl,
+        state: 'home',
         inject_data: {
             mapbox_access_token: config('mapbox-api').accessToken
         },
@@ -112,7 +108,7 @@ router.get('/user-settings', function(req, res) {
             state: ''
         });
     } else {
-        res.redirect('/home');
+        res.redirect('/');
     }
 });