+++ scripts/db/014-projects-table-add-image-column.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `arbimon2`.`projects`
+ADD COLUMN `image` VARCHAR(255) DEFAULT NULL;