+++ app/model/projects.js
@@ -93,7 +93,8 @@ var Projects = {
             whereExp.push("(p.name LIKE '%"+query.q+"%' OR p.description LIKE '%"+query.q+"%')");
         }
         if(query.hasOwnProperty('featured')) {
-            whereExp.push("p.featured = 1");
+            selectExtra += 'featured, image, '
+            whereExp.push("p.featured = 1 OR p.featured = 2");
         }
 
         if(!whereExp.length) {