+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Resolved issues:
 
 New features:
 
+- CE-504 Revamp Arbimon homepage
+- CE-742 The user can only add users to a project who are RFCx-connected
+- CE-763 Add info to top highlighted project cards
 - CE-742 User can only add users to a project who are RFCx-connected
 
 Resolved issues: